Improvement of Trigger
Changed the trigger to now also be able to count the number of prohibited keywords mentioned in the case description as well as collecting the mentioned words in a list. Both the counter and list are then displayed to the user in the Description field on the created child case.