Bug fixing
Trigger was failing when no ice cream flavour was selected as the list was null. Therefore, both the splitting of the selected flavours did not work as well as the size() command to count the elements in the list (both failing due to null pointer exceptions). This has now been fixed by adding the appropriate null-checks.