Added new Trigger
This trigger creates a Comparable Opportunity junction record as detail of a new opporunity for each past opporutnity that is similar to the new one in certain aspects. Thereby, the user is able to see a list of similar opportunities on their newly created one to potentially ask the Owners of those for advice on closing.